ANDROID: cpufreq: Add transient max freq for hw thermals
Allows thermal max freq notification to the scheduler without
interfacing it with the thermal driver.  Adds a new "transient"
field in the cpufreq policy to be notified by thermals and consumed
by the scheduler.  Note that this could be expanded for other
notifiers and consumers, in this CL though the intent is
only to allow thermals including a HW engine to notify
a transient value to be used by the scheduler but not by
the governor.

Signed-off-by: Andres Oportus <andresoportus@google.com>
Bug: 37716228
Test: Check that under thermal throttling the scheduler reports
less capacity in use as in:
2017-06-13 17:23:05>as 'dmesg|grep CPU|tail -1'
[  449.193273] CPU7: update max cpu_capacity 984

Change-Id: I13ddd3d62db765dc3b47ccaddeacfd92f79c5a2a
Signed-off-by: Haseo97 <jawgattus@gmail.com>